backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
openai==1.3.0
transformers==4.35.2
torch==2.1.1
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
requests==2.31.0
pytest==7.4.3
httpx==0.25.2 